requirements.txt
boto3
Django==4.2.1
django-autoslug==1.9.9
django-storages==1.13.2
djangorestframework==3.14.0
django-cors-headers
djangorestframework-simplejwt
mock==5.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
gunicorn
